BACKPORT: kbuild: Fix clang detection
We cannot detect clang before including the arch Makefile, because that
can set the default cross compiler. We also cannot detect clang after
including the arch Makefile, because powerpc wants to know about clang.
Solve this by using an deferred variable. This costs us a few shell
invocations, but this is only a constant number.

Reported-by: Behan Webster <behanw@converseincode.com>
Reported-by: Anton Blanchard <anton@samba.org>
Signed-off-by: Michal Marek <mmarek@suse.com>
Link: https://git.kernel.org/linus/5631d9c429857194bd55d7bcd8fa5bdd1a9899a3
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
Change-Id: Ie688f0f3f142de3f95c93394636c1993761fcd3d